worker-service-aws/src/requirements.txt
gunicorn==23.0.0
redis==6.4.0
boto3==1.40.63
botocore==1.40.63
python-json-logger==3.3.0
psycopg2-binary==2.9.10